Adding Buttons to Create and Clear Habits 1. Added button to create initial habits, that triggers a POST request to `/create_initial_habits` 2. Added button to clear all habits from table, which deletes the habits table and recreates it as an empty table. This button triggers a POST request to `/clear_habits_table`
